requirements.txt
google-cloud-bigquery
fastapi
pydantic
uvicorn
pandas
pyarrow
pytest
httpx
python-dotenv
